% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blended1.R
\name{blended1}
\alias{blended1}
\title{\code{BlendedEqn1}}
\usage{
blended1(
  i,
  n2up,
  NOther,
  NFix,
  z,
  light,
  DN,
  tt,
  tempup,
  K600mean,
  gas,
  n,
  nsatup,
  nsatdown,
  lag
)
}
\arguments{
\item{name}{description}
}
\value{
description here
}
\description{
Internal function.
}
