% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blended2.R
\name{blended2}
\alias{blended2}
\title{\code{BlendedEqn2}}
\usage{
blended2(
  i,
  n2up,
  NConsume,
  NFix,
  z,
  light,
  DN,
  tt,
  tempup,
  K600mean,
  gas,
  n,
  nsatup,
  nsatdown,
  lag
)
}
\arguments{
\item{name}{description}
}
\value{
description here
}
\description{
Internal function.
}
