% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blended3.R
\name{blended3}
\alias{blended3}
\title{\code{BlendedEqn3}}
\usage{
blended3(
  i,
  n2up,
  NOther,
  NFix,
  z,
  light,
  DN,
  tt,
  tempup,
  K600mean,
  gas,
  n,
  nsatup,
  nsatdown,
  lag
)
}
\arguments{
\item{name}{description}
}
\value{
description here
}
\description{
Internal function.
}
