% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_NEON.R
\name{clean_NEON}
\alias{clean_NEON}
\title{Clean NEON metabolism parameter data and prep for two-station stream metabolism modeling}
\usage{
clean_NEON(data, k600_clean, k600_fit)
}
\arguments{
\item{data}{Dataframe of metabolism parameter time series, \code{data} in list output of the @request_NEON function}

\item{k600_clean}{Dataframe of summarized K600 estimates and parameters used for K600 calculations, \code{k600_clean} in list output of the @request_NEON function}

\item{k600_fit}{Linear model output for the relationship between mean Q and K600 at the site, \code{k600_fit} in list output of the @request_NEON function}
}
\value{

}
\description{
Clean NEON metabolism parameter data and prep for two-station stream metabolism modeling
}
\examples{

}
\seealso{

}
