% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_date_to_doyhr.R
\name{convert_UTC_to_solartime}
\alias{convert_UTC_to_solartime}
\title{Convert DateTime from UTC to local solar time}
\usage{
convert_UTC_to_solartime(date.time, longitude, time.type = "mean solar")
}
\arguments{
\item{date.time}{date-time values in POSIXct format and UTC timezone.}

\item{longitude}{numeric, in degrees, either positive and unitted ("degE" or
"degW") or with sign indicating direction (positive = East)}

\item{time.type}{character. "apparent solar", i.e. true solar time, is noon
when the sun is at its zenith. "mean solar" approximates apparent solar
time but with noons exactly 24 hours apart. Elsewhere in this package,
variables named "solar.time" are mean solar time, whereas "app.solar.time"
is apparent solar and "any.solar.time" is either.}
}
\value{
a POSIXct object that says it's in tz="UTC" but that's actually in
solar time, with noon being very close to solar noon
}
\description{
Convert DateTime from UTC to local solar time, which may be either apparent
solar (perfect match between noon and solar zenith) or mean solar (exactly 24
hours between solar noons). This script is forked from the \code{streamMetabolizer}
package (with some changes), and is reproduced here (instead of linking the
package as a dependency) for forward compatability, as \code{streamMetabolizer}
is no longer under active development, and dependencies \code{unitted} and
\code{LakeMetabolizer} are not available for R versions >= 4.0.
}
\references{
Yard, Bennett, Mietz, Coggins, Stevens, Hueftle, and Blinn. 2005.
Influence of topographic complexity on solar insolation estimates for the
Colorado River, Grand Canyon, AZ. Ecological Modelling.
}
