% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_twostation.R
\name{fit_twostation}
\alias{fit_twostation}
\title{Model two-station stream metabolism}
\usage{
fit_twostation(data, k600_fit)
}
\arguments{
\item{data}{Dataframe of raw data as returned in the first object from @request_NEON 's output list}

\item{k600_fit}{Linear model output for the relationship between mean Q and K600 at the site, the third object returned by @request_NEON}
}
\value{

}
\description{
Model two-station stream metabolism
}
\examples{

}
\seealso{

}
