% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Kcor.R
\name{Kcor}
\alias{Kcor}
\title{\code{Kcor} Estimate K, given water temperature and K600 values}
\usage{
Kcor(temp, K600)
}
\arguments{
\item{temp}{Water temperature in degrees C}

\item{K600}{Normalized K600 reaeration rate}
}
\value{
Populate here
}
\description{
Internal function. See Wanninkhof 1992 and Hall et al 2016 supplemental material for more detailed information.
}
\references{
Populate here
}
