% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lightIndependent.R
\name{lightIndependent}
\alias{lightIndependent}
\title{\code{lightIndependent}}
\usage{
lightIndependent(
  i,
  n2up,
  NConsume,
  z,
  DN,
  tt,
  tempup,
  K600mean,
  gas,
  n,
  nsatup,
  nsatdown
)
}
\arguments{
\item{name}{description}
}
\value{
description here
}
\description{
Internal function.
}
