% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/N2TimeSeries.R
\name{N2TimeSeries}
\alias{N2TimeSeries}
\title{\code{N2TimeSeries} Return modeled oxygen time series from median GPP, ER estimates}
\usage{
N2TimeSeries(
  NConsume = NULL,
  DN,
  timeup,
  timedown,
  NFix = NULL,
  NOther = NULL,
  n2up,
  z,
  light,
  tt,
  tempup,
  K600mean,
  gas,
  n,
  nsatup,
  nsatdown,
  lag,
  n2down,
  eqn
)
}
\arguments{
\item{NConsume}{Description}

\item{DN}{Description}

\item{z}{Description}

\item{tt}{Description}

\item{K600mean}{Description}

\item{data}{Dataframe of cleaned raw two station data (ex. "TS_S1S2")}

\item{upName}{Description}

\item{downName}{Description}
}
\value{
Populate here
}
\description{
Internal function.
}
\references{
Populate here
}
