% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nifong.R
\name{nifong}
\alias{nifong}
\title{\code{NifongEqn} Model N2 using two-station equation from Nifong et al. 2020}
\usage{
nifong(
  i,
  n2up,
  NConsume,
  z,
  light,
  DN,
  tt,
  tempup,
  K600mean,
  gas,
  n,
  nsatup,
  nsatdown,
  lag
)
}
\arguments{
\item{name}{description}
}
\value{
description here
}
\description{
Internal function.
}
