% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/O2_twoStation.R
\name{O2_twoStation}
\alias{O2_twoStation}
\title{\code{2stationO2}}
\usage{
O2_twoStation(
  i,
  oxyup,
  GPP,
  z,
  light,
  ER,
  tt,
  tempup,
  K600mean,
  gas,
  n,
  osatup,
  osatdown,
  lag
)
}
\arguments{
\item{name}{description}
}
\value{
description here
}
\description{
Internal function.
}
