% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/O2TimeSeries.R
\name{O2TimeSeries}
\alias{O2TimeSeries}
\title{O2TimeSeries FUNCTION: Return modeled oxygen time series from median GPP,
ER estimates. Internal function.}
\usage{
O2TimeSeries(GPP, ER, O2data, Kmean, z, tt, upName, downName)
}
\description{
ARGUMENTS:
GPP
ER
}
