% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reisinger.R
\name{reisinger}
\alias{reisinger}
\title{\code{ReisingerEqn} Model N2 using two-station equation from reisinger et al 2016}
\usage{
reisinger(i, n2up, z, DN, tt, tempup, K600mean, gas, n, nsatup, nsatdown, lag)
}
\arguments{
\item{name}{description}
}
\value{
description here
}
\description{
Internal function.
}
