% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request_NEON.R
\name{request_NEON}
\alias{request_NEON}
\title{Pull data products necessary for two-station metabolism modeling at NEON sites}
\usage{
request_NEON(NEONsites, startdate, enddate)
}
\arguments{
\item{NEONsites}{character string specifying 4-letter NEON site code to
request data from (ex. \code{"HOPB"}). Can be more than one site
(ex.\code{c("HOPB", "BLDE")} but be warned data pull will take longer)}

\item{startdate}{YYYY-MM character string defining start month and year for
data request}

\item{enddate}{YYY-MM character string defining end month and year for
data request}
}
\value{
List of four dataframes, \code{data}, \code{k600_clean}, \code{k600_fit}, and
\code{k600_expanded}. NOTE: Explain in detail what these dataframes contain.
}
\description{
Pull data products necessary for two-station metabolism modeling at NEON sites
}
\examples{
\dontrun{
data <- request_NEON(NEONsites = "HOPB", startdate = "2018-01", enddate = "2018-12")
}
}
