% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tspost.R
\name{tspost}
\alias{tspost}
\title{tspost FUNCTION: Calculate posterior probablility of model
This function calculates the posterior probability of the model given
parameters. Internal function, called within twostationpostsum}
\usage{
tspost(
  MET,
  tempup,
  tempdown,
  oxyup,
  oxydown,
  light,
  tt,
  z,
  osat,
  Kmean,
  Ksd,
  ...
)
}
\description{
ARGUMENTS:
MET          Dataframe name of cleaned raw two station data (ex. "TS_S1S2")
tempup        Temperature data from upstream station, deg C
tempdown      Temperature data from downstream station, deg C
oxyup         Oxygen data from upstream station, mg-O2/L
oxydown       Oxygen data from downstream station, mg-O2/L
Light         any light unit
tt            travel time, days
Kmean
Ksd
}
