% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tspost_N2.R
\name{tspost_N2}
\alias{tspost_N2}
\title{\code{tspost} Calculate posterior probablility of two-station metabolism model.}
\usage{
tspost_N2(
  MET,
  tempup,
  tempdown,
  n2up,
  n2down,
  light,
  tt,
  z,
  nsatup,
  nsatdown,
  K600mean,
  K600sd,
  gas,
  n,
  eqn,
  lag
)
}
\arguments{
\item{MET}{Dataframe name of cleaned raw two station data (ex. "TS_S1S2")}

\item{tempup}{Temperature data from upstream station, deg C}

\item{tempdown}{Temperature data from downstream station, deg C}

\item{n2up}{Oxygen data from upstream station, mg-O2/L}

\item{n2down}{Oxygen data from downstream station, mg-O2/L}

\item{light}{any light unit}

\item{tt}{travel time, days}

\item{z}{Description}

\item{nsatup}{Description}

\item{nsatdown}{Description}

\item{Kmean}{Description}

\item{Ksd}{Description}
}
\value{
Populate here
}
\description{
This function calculates the posterior probability of the the two station metabolism model given
parameters. Internal function, called within twostationpostsum
}
\references{
Populate here
}
