% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twostationpostsum.R
\name{twostationpostsum}
\alias{twostationpostsum}
\title{\code{twostationpostsum} Calculate two station MCMC or Bayes during one day.}
\usage{
twostationpostsum(
  data,
  upName,
  downName,
  start,
  z,
  tt,
  K600mean,
  K600sd,
  nbatch,
  scale,
  modType,
  gas,
  n,
  eqn
)
}
\arguments{
\item{data}{Dataframe containing formatted raw two station data, as returned by request_NEON in $data}

\item{upName}{Character string denoting name of upstream station (ex. "S1")}

\item{downName}{Character string denoting name of downstream station (ex. "S2")}

\item{start}{Description}

\item{z}{Numeric, average depth on day of modeling}

\item{tt}{Numeric, travel time between stations on day of modeling}

\item{K600mean}{Mean K600 for informative prior}

\item{K600sd}{Standard deviation of K600 for informative prior}

\item{nbatch}{Number of MCMC trials}

\item{scale}{Description}

\item{modType}{Description}

\item{gas}{Either "O2" or "N2" denoting gas type}
}
\value{
Populate here
}
\description{
Internal function. Function runs the MCMC and returns posterior distributions
}
\references{
Populate here
}
