% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_twostation.R
\name{fit_twostation}
\alias{fit_twostation}
\title{Model two-station stream metabolism}
\usage{
fit_twostation(data)
}
\arguments{
\item{data}{Dataframe containing formatted raw two station data, as
returned by @request_NEON in $data}
}
\value{

}
\description{
Model two-station stream metabolism
}
\examples{

}
\seealso{

}
