% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_twostation.R
\name{fit_twostation}
\alias{fit_twostation}
\title{Model two-station stream metabolism}
\usage{
fit_twostation(data, modType, nbatch = 1e+05)
}
\arguments{
\item{data}{The output from @request_NEON function, a list of 4 objects: \code{data}: dataframe containing raw site data, \code{k600_clean}: Dataframe of summarized K600 estimates and parameters used for calculations,\code{k600_fit}: Linear model output for the relationship between mean Q and K600 at the site, \code{k600_expanded}: dataframe of all data used in K600 calculations}

\item{modType}{Model type as a character string. Either "mle" for maximum likelyhood estimation or "bayes" for Bayesian estimation.}

\item{nbatch}{Numeric, number of batches for Bayesian chains. Default is 100,000.}
}
\value{

}
\description{
Model two-station stream metabolism
}
\examples{

}
\seealso{

}
