% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Kcor.R
\name{Kcor}
\alias{Kcor}
\title{Estimate K at temperature, given temperature and
normalized K600 values. Internal function. See Wanninkhof 1992 and
Hall et al 2016 supplemental material}
\usage{
Kcor(temp, K600)
}
\description{
Estimate K at temperature, given temperature and
normalized K600 values. Internal function. See Wanninkhof 1992 and
Hall et al 2016 supplemental material
}
