% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Kcor.R
\name{Kcor}
\alias{Kcor}
\title{\code{Kcor} Estimate kO2, given water temperature and K600 values}
\usage{
Kcor(temp, K600, gas, n = 0.5)
}
\arguments{
\item{temp}{Water temperature in degrees C}

\item{K600}{Normalized K600 reaeration rate}

\item{gas}{Either "O2" or "N2", used for calculating Schmidt number}

\item{n}{Default is 0.5 for wavy water, but 0.667 refers to still water, see Jahne 1987}
}
\value{
Populate here
}
\description{
Internal function. See Wanninkhof 1992 and Hall et al 2016 supplemental material for more detailed information.
}
\references{
Populate here
}
