% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/O2TimeSeries.R
\name{O2TimeSeries}
\alias{O2TimeSeries}
\title{\code{O2TimeSeries} Return modeled oxygen time series from median GPP, ER estimates}
\usage{
O2TimeSeries(GPP, ER, O2data, Kmean, z, tt, upName, downName)
}
\arguments{
\item{GPP}{Description}

\item{ER}{Description}

\item{O2data}{Dataframe of cleaned raw two station data (ex. "TS_S1S2")}

\item{Kmean}{Description}

\item{z}{Description}

\item{tt}{Description}

\item{upName}{Description}

\item{downName}{Description}
}
\value{
Populate here
}
\description{
Internal function.
}
\references{
Populate here
}
