% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twostationpostsum.R
\name{twostationpostsum}
\alias{twostationpostsum}
\title{Calculate two station MCMC during one day. Internal function. Function runs
the MCMC and returns posterior distributions}
\usage{
twostationpostsum(
  O2data,
  upName,
  downName,
  start,
  z,
  tt,
  Kmean,
  Ksd,
  nbatch,
  scale
)
}
\description{
Calculate two station MCMC during one day. Internal function. Function runs
the MCMC and returns posterior distributions
}
