% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twostationpostsum.R
\name{twostationpostsum}
\alias{twostationpostsum}
\title{\code{twostationpostsum} Calculate two station MCMC or Bayes during one day.}
\usage{
twostationpostsum(
  O2data,
  upName,
  downName,
  start,
  z,
  tt,
  Kmean,
  Ksd,
  nbatch,
  scale,
  modType
)
}
\arguments{
\item{O2data}{Dataframe containing formatted raw two station data, as returned by request_NEON in $data}

\item{upName}{Character string denoting name of upstream station (ex. "S1")}

\item{downName}{Character string denoting name of downstream station (ex. "S2")}

\item{start}{Description}

\item{z}{Numeric, average depth on day of modeling}

\item{tt}{Numeric, travel time between stations on day of modeling}

\item{Kmean}{Mean K for informative prior}

\item{Ksd}{Standard deviation of K for informative prior}

\item{nbatch}{Number of MCMC trials}

\item{scale}{Description}

\item{modType}{Description}
}
\value{
Populate here
}
\description{
Internal function. Function runs the MCMC and returns posterior distributions
}
\references{
Populate here
}
