% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twostationpostsum.R
\name{twostationpostsum}
\alias{twostationpostsum}
\title{Calculate two station MCMC during one day. Internal function. Function runs
the MCMC and returns posterior distributions}
\usage{
twostationpostsum(
  O2data,
  upName,
  downName,
  start,
  z,
  tt,
  Kmean,
  Ksd,
  nbatch,
  scale,
  modType
)
}
\description{
O2data    Dataframe containing formatted raw two station data, as
returned by request_NEON in $data
upName    Character string denoting name of upstream station (ex. "S1")
downName  Character string denoting name of downstream station (ex. "S2")
init      Initial guess for the range of GPP and |ER|
z         Numeric, average depth on day of modeling
tt        Numeric, travel time between stations on day of modeling
Kmean     Mean K for informative prior
Ksd       Standard deviation of K for informative prior
nbatch    Number of MCMC trials
scale
}
